+++ common/src/main/kotlin/com/cobblemon/mod/common/api/spawning/spawner/PlayerSpawner.kt
@@ -20,6 +20,7 @@ import kotlin.math.cos
 import kotlin.math.sin
 import kotlin.random.Random
 import net.minecraft.server.network.ServerPlayerEntity
+import net.minecraft.server.world.ServerWorld
 import net.minecraft.util.math.MathHelper.PI
 import net.minecraft.util.math.MathHelper.ceil
 
@@ -57,7 +58,7 @@ class PlayerSpawner(player: ServerPlayerEntity, spawns: SpawnPool, manager: Spaw
 
         return SpawningArea(
             cause = cause,
-            world = player.world,
+            world = player.world as ServerWorld,
             baseX = ceil(x - sliceDiameter / 2F),
             baseY = ceil(center.y - sliceHeight / 2F),
             baseZ = ceil(z - sliceDiameter / 2F),